Responsive text
We had some issues, where text on mobile would be displayed in a too
large font-size. Upon further investigation, and in coordination with
our R&D department, we have come to the conclusion that using media
queries is an appropriate solution to this problem.

As an initial fix, this one-step screen size query should resolve most
issues.